fl/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "src"
version = "1.0.0"
description = "Federated Learning with PyTorch and authenticated Flower "
license = "Apache-2.0"
dependencies = [
    "flwr>=1.13.1",
    "flwr-datasets[vision]>=0.4.0",
    "torch>=2.5.0,<3.0.0",
    "torchvision>=0.20.1,<0.21.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.flwr.app]
publisher = "flwrlabs"

[tool.flwr.app.components]
serverapp = "src.server_app:app"
clientapp = "src.client_app:app"

[tool.flwr.app.config]
num-server-rounds = 3
fraction-evaluate = 0.5
local-epochs = 1
learning-rate = 0.1
batch-size = 32

[tool.flwr.federations]
default = "my-federation"

[tool.flwr.federations.my-federation]
address = "127.0.0.1:9093"                # Address of the Exec API
root-certificates = "certificates/ca.crt"
